#8 Order matching to be executed by parts
Orders are not matched by parts, which means that the transaction covers
only the part in which an order is partially matched against the
available size of the matching order found, instead of covering the
whole matching process (until the order is depleted). This saves failing
transactions, allowing for more granularity and less failure.