Modified interface to match-once and order matching.
Changed location of cross-order execution function to the engine, no need for it to be outside, just the on-execution and on-cancellation functions must remain outside since they have the information details to be done upon these operations.

Tests adapted to these changes. Removed bulk test.

Has to do with accommodating order cancellation of Issue #27